PCB/requirements.txt
aiofiles==25.1.0
blinker==1.9.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Flask==3.1.2
fonttools==4.61.0
gerbonara==1.6.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
Hypercorn==0.18.0
hyperframe==6.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pip==25.2
priority==2.0.0
pyparsing==3.2.5
pypdf==6.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
Quart==0.20.0
reportlab==4.4.5
rtree==1.4.1
six==1.17.0
tzdata==2025.2
Werkzeug==3.1.4
wsproto==1.3.2
